refactor: move some time and currency consts into constants.rs
